diffusion_visualizer/requirements.txt
gradio>=4.0.0
torch>=2.0.0
diffusers>=0.25.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
